:truck: Move the atmosphere preprocessor to the plugin
