+++ CMakeLists.txt
@@ -324,5 +324,4 @@ add_subdirectory(config)
 add_subdirectory(resources)
 add_subdirectory(src)
 add_subdirectory(plugins)
-add_subdirectory(tools/atmosphere-generator)
 add_subdirectory(tools/eclipse-shadow-generator)
