+++ plugins/csp-atmospheres/atmosphere-generator/angstromMode.cpp
