+++ plugins/csp-atmospheres/atmosphere-generator/angstromMode.hpp
