+++ plugins/csp-atmospheres/atmosphere-generator/bhmie.cpp
