+++ plugins/csp-atmospheres/atmosphere-generator/density-settings/earth_bruneton_aerosols.json
