+++ plugins/csp-atmospheres/atmosphere-generator/density-settings/earth_bruneton_molecules.json
