+++ plugins/csp-atmospheres/atmosphere-generator/density-settings/earth_rain.json
