+++ plugins/csp-atmospheres/atmosphere-generator/density-settings/mars_collienne_aerosols.json
