+++ plugins/csp-atmospheres/atmosphere-generator/density-settings/mars_cosmoscout_molecules.json
