+++ plugins/csp-atmospheres/atmosphere-generator/density-settings/mars_costa_aerosols.json
