+++ plugins/csp-atmospheres/atmosphere-generator/densityMode.hpp
