+++ plugins/csp-atmospheres/atmosphere-generator/hulstMode.cpp
