+++ plugins/csp-atmospheres/atmosphere-generator/hulstMode.hpp
