+++ plugins/csp-atmospheres/atmosphere-generator/manualMode.cpp
