+++ plugins/csp-atmospheres/atmosphere-generator/manualMode.hpp
