+++ plugins/csp-atmospheres/atmosphere-generator/mie-settings/earth_haze.json
