+++ plugins/csp-atmospheres/atmosphere-generator/mie-settings/earth_rain.json
