+++ plugins/csp-atmospheres/atmosphere-generator/mie-settings/mars_ehlers.json
