+++ plugins/csp-atmospheres/atmosphere-generator/ozoneMode.cpp
@@ -7,7 +7,7 @@
 
 #include "ozoneMode.hpp"
 
-#include "../../src/cs-utils/utils.hpp"
+#include "../../../src/cs-utils/utils.hpp"
 #include "common.hpp"
 
 #include <glm/gtc/constants.hpp>
