+++ plugins/csp-atmospheres/atmosphere-generator/ozoneMode.hpp
