+++ plugins/csp-atmospheres/atmosphere-generator/parametricModes.cpp
@@ -9,7 +9,7 @@
 
 #include "common.hpp"
 
-#include "../../src/cs-utils/utils.hpp"
+#include "../../../src/cs-utils/utils.hpp"
 
 #include <glm/gtc/constants.hpp>
 #include <nlohmann/json.hpp>
