+++ plugins/csp-atmospheres/atmosphere-generator/parametricModes.hpp
