+++ plugins/csp-atmospheres/atmosphere-generator/rayleighMode.cpp
