+++ plugins/csp-atmospheres/atmosphere-generator/rayleighMode.hpp
